requirements.txt
django>=2.1
psycopg2
djangorestframework
djangorestframework_simplejwt
Pillow
django-cors-headers